requirements-remote.txt
# Remote MCP Server - Minimal dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
